Types incorrectly marked as @internal
Types only need to be marked as `@internal` if they would be exported
from the top level `index.ts` file, but you don't want them to. Marking
something as `@internal` means they are stripped from the compiled type
files and this causes type errors because some files will try to
reference types that do not exist anymore.

Example error when running `tsc` in a project where `@pulumi/cdk` is
installed.

```
node_modules/@pulumi/cdk/lib/converters/app-converter.d.ts:6:10 - error TS2305: Module '"../interop"' has no exported member 'CdkConstruct'.

6 import { CdkConstruct } from '../interop';
           ~~~~~~~~~~~~
```

I've unmarked things as `@internal` that are not exported and moved some
things to a new `interna/interop.ts` file that were in an exported file.
Now that they are not in an exported file we don't have to mark them as
`@internal`
